ref: ssh web connection structs moved to different file. error handled in tunnel close
